# Speedup: PrettyPrint, Render, and Parse