@matthinrichsen-wf 2023-08-02
@xuri I believe I've addressed the formatting issue - and as far as I can tell the code coverage should now be at 100% (using `go test -v -race -coverprofile=coverage.txt -covermode=atomic ./...`)